+++ k256/src/arithmetic/field.rs
@@ -38,6 +38,7 @@ use core::{
 };
 use elliptic_curve::{
     ff::{self, Field, PrimeField},
+    ops::Invert,
     rand_core::RngCore,
     subtle::{Choice, ConditionallySelectable, ConstantTimeEq, CtOption},
     zeroize::DefaultIsZeroes,
@@ -245,6 +246,14 @@ impl FieldElement {
     }
 }
 
+impl Invert for FieldElement {
+    type Output = CtOption<Self>;
+
+    fn invert(&self) -> CtOption<Self> {
+        self.invert()
+    }
+}
+
 impl Field for FieldElement {
     const ZERO: Self = Self::ZERO;
     const ONE: Self = Self::ONE;
@@ -432,6 +441,7 @@ impl Mul<FieldElement> for FieldElement {
 impl Mul<&FieldElement> for FieldElement {
     type Output = FieldElement;
 
+    #[inline(always)]
     fn mul(self, other: &FieldElement) -> FieldElement {
         FieldElement(self.0.mul(&(other.0)))
     }
@@ -480,6 +490,7 @@ impl Sum for FieldElement {
 }
 
 impl<'a> Sum<&'a FieldElement> for FieldElement {
+    #[inline]
     fn sum<I: Iterator<Item = &'a FieldElement>>(iter: I) -> Self {
         iter.copied().sum()
     }
@@ -500,8 +511,10 @@ impl<'a> Product<&'a FieldElement> for FieldElement {
 #[cfg(test)]
 mod tests {
     use elliptic_curve::ff::{Field, PrimeField};
+    use elliptic_curve::ops::BatchInvert;
     use num_bigint::{BigUint, ToBigUint};
     use proptest::prelude::*;
+    use rand_core::OsRng;
 
     use super::FieldElement;
     use crate::{
@@ -509,6 +522,9 @@ mod tests {
         test_vectors::field::DBL_TEST_VECTORS,
         FieldBytes,
     };
+
+    #[cfg(feature = "alloc")]
+    use alloc::vec::Vec;
 
     impl From<&BigUint> for FieldElement {
         fn from(x: &BigUint) -> Self {
@@ -670,6 +686,31 @@ mod tests {
         let two = one + &one;
         let inv_two = two.invert().unwrap();
         assert_eq!((two * &inv_two).normalize(), one);
+    }
+
+    #[test]
+    fn batch_invert_array() {
+        let k: FieldElement = FieldElement::random(&mut OsRng);
+        let l: FieldElement = FieldElement::random(&mut OsRng);
+
+        let expected = [k.invert().unwrap(), l.invert().unwrap()];
+        assert_eq!(
+            <FieldElement as BatchInvert<_>>::batch_invert(&[k, l]).unwrap(),
+            expected
+        );
+    }
+
+    #[test]
+    #[cfg(feature = "alloc")]
+    fn batch_invert() {
+        let k: FieldElement = FieldElement::random(&mut OsRng);
+        let l: FieldElement = FieldElement::random(&mut OsRng);
+
+        let expected = vec![k.invert().unwrap(), l.invert().unwrap()];
+        let field_elements = vec![k, l];
+        let res: Vec<_> =
+            <FieldElement as BatchInvert<_>>::batch_invert(field_elements.as_slice()).unwrap();
+        assert_eq!(res, expected);
     }
 
     #[test]