+++ p521/CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.13.3 (2023-11-11)
+### Added
+- Implement hash2curve ([#964])
+
+### Fixed
+- Panics when decoding `FieldElement`s ([#967])
+
+[#964]: https://github.com/RustCrypto/elliptic-curves/pull/964
+[#967]: https://github.com/RustCrypto/elliptic-curves/pull/967
+
 ## 0.13.2 (2023-11-09)
 ### Added
 - `serde` feature ([#962])